Copy .keep files without the .keep extension
Sometimes it can be useful to copy .yas files as-is, for example when making
ptemplate template templates. In this case, they can be named <name>.yas.keep,
resulting in the file <name>.yas appearing as a result. Note that this feature
is incomplete yet, as .keep files named like non-.keep files are copied too,
instead of a warning being issued.