Abstract template prompting system (support ivy, ...)
Abstract away template prompting in a way that does not depend on helm. Remove
helm-ptemplate-new-project; ptemplate-expand-template is now interactive
instead, and asks the user for a template with their completion system of
choice, controlled by the variable `ptemplate-template-prompt-function'. Update
README.

Add a "Configuration" section describing the various configurations of
ptemplate.

Add a variable, `ptemplate-default-workspace', which configures the default
workspace used if the selected template's type is not associated with any
workspace in `ptemplate-workspace-alist'.