Remove limitations: allow multiple templates to be expanded at once
To accomplish this, the snippet chain, source and target file variables had to
be made buffer local:

The first variable represented a problem: it was a list that had to be shared
between multiple buffers. The solution was to abuse conses as pointers, their
cdr pointing to the actual shared list. The conses wouldn't be shared, but they
all would alter the same list (with nconc and pop operations).

The latter two are inherited from the context of
`ptemplate--snippet-chain-contine` (which mostly is form previous snippet chain
buffers except when ptemplate-expand-snippet is called, which let-binds them)
and setq-local'd to the target.

The result is that any number of snippet expansions may take place at any time,
concurrently, so update the README.md to remove the # Limitations section.