ptemplate!: add a :finalize block (after expansion)
Add a :finalize block, which is executed after all snippets have been expanded
and all files copied.

Refactor the snippet chain subsystem and extend it with a
`ptemplate--snippet-chain-inherit' variable, which is a list of symbols passed
on to all buffers therein. This way, that subsystem won't become a tangled mess
referencing variables all around.

Also extend it with a `ptemplate--snippet-chain-finalize-hook`, run in
`ptemplate--snippet-chain-continue` (and consequently in
`ptemplate-snippet-chain-next`, etc...) when the next item in the snippet chain
is nil. This way, :finalize can be implemented.

Breaking change: ptemplate!: :after becomes :after-copy.