bugfix: snippet-chain environment with multiple snippets
Previously, multiple .yas snippets did not work with snippet chain environments.
Fix this by refactoring the snippet chain subsystem to use an alist of (SYMBOL .
VALUE) instead of a list of SYMBOLS to inherit (`ptemplate--snippet-chain-env').

The snippet-chain variables (`ptemplate--snippet-chain' et al.) are still
inherited using (buffer-local-value oldbuf), but that seems to work.