Support remapping template files
`ptemplate-expand-template' was rearchitected to now not only store paths and
assume them to be the same in the target, but to now also store the target of
each template file. This allows for mapping and remapping files to different
targets. Add some new utility functions to allow making use of this new feature.