bugfix: run `ptemplate--chain-finalize-hook' at the very end
`ptemplate--snippet-chain-hook' was previously run in
`ptemplate--snippet-chain-next' *before* killing the buffer. This caused issues
when using `find-file' to find the last snippet chain file in `ptemplate!''s
:finalize.

`ptemplate--finalize-hook' was not run if there were no snippets, because it was
passed to `ptemplate--snippet-chain-start', only run if there are snippets. Fix
that, running it "manually" if there are no snippets instead.