Fix inherit and include, :open last file in fg
Inherit and include now actually work (with test cases). To do this,
each source file in the file map needs to know from where it is
imported, which was done by making the source a cons (PREFIX . RSRC)
instead.

`ptemplate!''s :open now only opens the last file in the foreground, all
others in the background. Doing it the other way is redundant anyway, as
`find-file' always opens in the foreground, and so opening all others in
the foreground would have the same effect, with more
flickering (`find-file's overwrite each other).

Improve code.