Support nil maps properly
As previously implemented, but wrongly documented, overriding is done by
target, not by source (if two targets are equal, one wins, not if the
sources are equal). This is better, since it makes more sense.

Because overriding is now done by target, nil maps are now
mappings *from* nil, not to nil. (The latter is only useful with
`ptemplate-remap', in which case it would be equivalent to
`ptemplate-ignore').

Add a test case for nil maps and improve the file names of the others.