Fix `ptemplate-inherit' with multiple SRCs
Previously, `ptemplate-inherit' and `ptemplate-include', when used with
multiple SRCs, would not override each individual inherited source. This
was due to a change in the behaviour of `ptemplate--override-files',
which made it not remove duplicates in a single file-map list. Refactor
that function to take an arbitrary number of arguments, powered by
`ptemplate--merge-filemaps'.

Remove "\(\)" except at the start of lines in docstrings. I
misunderstood an advice recommending to use such escapes: these are only
relevant at the beginning of lines, where, due to an Emacs bug in
versions below 27.1, they would cause built-in functions to incorrectly
parse s-expressions.