ptemplate-test.el: test template prompt functions
Add a new test, `ptemplate-prompt' (based on its previous, commented
form) which serves as a smoke-test for the template prompting
functionality. Depend on with-simulated-input for it.

ptemplate.el: `ptemplate-template-prompt-function' may now return nil to
indicate that the user didn't select anything, and the completing-read
based one does now (needed for the test). Because of that feature, bump
the minor version.