`ptemplate-map': support CONTENT
`ptemplate-map' now supports specifying a string to be used as content
for the source, instead of it being read from a file. This is needed for
generating C++ classes, as they need to be dynamically computed.
Implement this by extending `ptemplate--file-mapping' to have a content
field. `ptemplate--snippet-mapping' no longer takes a SRC file, but a
string instead. If truly large templates are needed, it may be extended
to do the fetching lazily.

Add a new `ptemplate--file-mapping' :type, `:mkdir', which allows
creating directories, ignoring SRC.