Use `error' instead of `lwarn'
Often, `lwarn' was used to notify the user of some error (duplicate
mappings, not a snippet, ...). This makes debugging harder, because
`debug-on-error' would not cause a backtrace to be shown. Since these
conditions are always wrong, use `error' instead.

Using `error' also simplifies the code and test-suite, allowing
`ptemplate-test--with-fatal-warnings' to be dropped.