`ptemplate--prompt-target': support workspace list
`ptemplate-default-workspace' and the associated workspaces of
`ptemplate-workspace-alist' may now be lists of workspaces in addition
to just workspace paths, in which case one will have to be chosen in
`ptemplate--prompt-target' first.

This is useful if some language needs to be used in multiple
contexts (work, personal projects, ...) that have separate workspace
folders.

Also change the `defcustom' :types of `ptemplate-default-workspace' and
`ptemplate-workspace-alist' to refer to `file's instead of strings.